refactor(FeatureGoogleAnalytics): rename namespace
